rename `virtual` module to `provider_index`
