+++ lib/spack/spack/provider_index.py
